Fix : "like and divide component"
